Moving new stream routing engine from sub- to superclass.
