--- graylog2-server/src/test/java/org/graylog2/streams/OtherStreamRouterEngineTest.java
@@ -1,29 +1,0 @@
-/**
- * This file is part of Graylog.
- *
- * Graylog is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Graylog is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.graylog2.streams;
-
-import org.graylog2.plugin.streams.Stream;
-
-import java.util.List;
-import java.util.concurrent.Executors;
-
-public class OtherStreamRouterEngineTest extends StreamRouterEngineTest {
-    @Override
-    protected StreamRouterEngine newEngine(List<Stream> streams) {
-        return new OtherStreamRouterEngine(streams, Executors.newSingleThreadExecutor(), streamFaultManager, streamMetrics);
-    }
-}
